requirements.txt
Django~=6.0.0
djangorestframework
django-cors-headers
Pillow